src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/resize-in-shadow-dom-expected.txt

CONSOLE ERROR: line 22: Uncaught TypeError: host.createShadowRoot is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
